konec velkeho uklidu na presenteru nastroju
