+++ app/templates/Sprava/nastroje.latte
@@ -19,22 +19,22 @@
     {/if}
     <a href="{link SpravaCc:prehledcc id=>null}" class="btn btn-default btn-sm" role="button"> Přehled platných čestných členství</a>
     <br><br>
-    <a href="{link Sprava:usersgraph id=>null}" class="btn btn-default btn-sm" role="button"> Graf uživatelů</a>
+    <a href="{link SpravaGrafu:usersgraph id=>null}" class="btn btn-default btn-sm" role="button"> Graf uživatelů</a>
     <br><br>
-    <a href="{link Sprava:ucty id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Aktuální stavy účtů</a>
-    <a href="{link Sprava:odchoziplatby id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Přehled odchozích plateb</a>
+    <a href="{link SpravaUctu:ucty id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Aktuální stavy účtů</a>
+    <a href="{link SpravaPlateb:odchoziplatby id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Přehled odchozích plateb</a>
     <br><br>
     
     {if $canSeePayments}
-        <a href="{link Sprava:platbycu id=>null}" class="btn btn-default btn-sm" role="button"> Platby z Č.Ú.</a>
-        <a href="{link Sprava:nesparovane id=>null}" class="btn btn-default btn-sm" role="button"> Nespárované platby</a>
+        <a href="{link SpravaPlateb:platbycu id=>null}" class="btn btn-default btn-sm" role="button"> Platby z Č.Ú.</a>
+        <a href="{link SpravaNesparovanych:nesparovane id=>null}" class="btn btn-default btn-sm" role="button"> Nespárované platby</a>
         <br><br>
-        <a href="{link Sprava:sms id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-phone"></i> Odeslat SMS správcům</a>
+        <a href="{link SpravaSms:sms id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-phone"></i> Odeslat SMS správcům</a>
         <br><br>
     {/if}
 
     {if $canSeeMailList}
-        <a href="{link Sprava:mailinglist id=>null}" class="btn btn-default btn-sm" role="button"> Mailing list</a>
+        <a href="{link SpravaMailu:mailinglist id=>null}" class="btn btn-default btn-sm" role="button"> Mailing list</a>
         <br><br>
     {/if}
