+++ app/templates/SpravaGrafu/usersgraph.latte
