+++ app/templates/Uzivatel/platba.latte
@@ -6,7 +6,7 @@
     <a href="{link Uzivatel:account id=>$u->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-arrow-up"></i> Zpět na přehled plateb</a>
     
     {if $canTransfer}
-    <a href="{link Sprava:prevod id=>$p->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Převést jinému uživateli (chybná platba)</a>
+    <a href="{link SpravaNesparovanych:prevod id=>$p->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Převést jinému uživateli (chybná platba)</a>
     {/if}
     {/ifset}
